+++ tools/archtests/source_packages_test.go
@@ -22,7 +22,7 @@ var grandfatheredSourcePackageLOCBudgets = map[string]int{
 	"github":          2102,
 	"googleworkspace": 827,
 	"grc":             1378,
-	"okta":            2361,
+	"okta":            2269,
 	"panopticon":      820,
 	"sentinelone":     2181,
 	"vulnview":        1064,
